move cursor state to its own hook
